+++ modules/python/test/tests_common.py
@@ -102,11 +102,12 @@ class NewOpenCVTests(unittest.TestCase):
 
 def intersectionRate(s1, s2):
 
+    # dtype is explicit: the geometry functions accept only CV_32S/CV_32F points.
     x1, y1, x2, y2 = s1
-    s1 = np.array([[x1, y1], [x2,y1], [x2, y2], [x1, y2]])
+    s1 = np.array([[x1, y1], [x2,y1], [x2, y2], [x1, y2]], dtype=np.int32)
 
     x1, y1, x2, y2 = s2
-    s2 = np.array([[x1, y1], [x2,y1], [x2, y2], [x1, y2]])
+    s2 = np.array([[x1, y1], [x2,y1], [x2, y2], [x1, y2]], dtype=np.int32)
 
     area, _intersection = cv.intersectConvexConvex(s1, s2)
     return 2 * area / (cv.contourArea(s1) + cv.contourArea(s2))
